+++ concept/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod layout;
+pub mod responsive;
 pub mod scroll;
 pub mod spring;
 pub mod svg_init;
