+++ concept/src/responsive/clickable.rs
@@ -1,5 +1,5 @@
-use super::get_normalize_scale;
-use super::Layout;
+use super::layout::get_normalize_scale;
+use super::layout::Layout;
 use guppies::glam::Mat4;
 use guppies::glam::Vec4;
